fix(@angular/cli):  Support XDG Base Directory Specfication
Reading throw  #16034 the angular config should be placed in the `$XDG_CONFIG_HOME/angular/` directory, but if the environmental variable $XDG_CONFIG_HOME is set the first check of `xdgConfigHome` function will only put it in the `$XDG_CONFIG_HOME` directory. Also renamed the config file from `.angular-config.json` to `config.json`   when it is in the `~/.config/angular` directory since it doesn't need to be hidden.